refactor(saga): Phase 2a move inbound event payloads to Contracts/
