+++ saga-microservice/Saga.Service/Contracts/Integration/InboundEvents/PaymentFailedEvent.cs
@@ -1,8 +1,8 @@
 using ECommerce.Shared.Infrastructure.EventBus;
 
-namespace Saga.Service.IntegrationEvents;
+namespace Saga.Service.Contracts.Integration.InboundEvents;
 
-public record PaymentVoidedEvent(
+public record PaymentFailedEvent(
     Guid PaymentId,
     Guid OrderId,
     string CustomerId,
