+++ saga-microservice/Saga.Service/Contracts/Integration/InboundEvents/PaymentVoidedEvent.cs
@@ -1,8 +1,8 @@
 using ECommerce.Shared.Infrastructure.EventBus;
 
-namespace Saga.Service.IntegrationEvents;
+namespace Saga.Service.Contracts.Integration.InboundEvents;
 
-public record PaymentFailedEvent(
+public record PaymentVoidedEvent(
     Guid PaymentId,
     Guid OrderId,
     string CustomerId,
