+++ saga-microservice/Saga.Service/Contracts/Integration/InboundEvents/ShipmentFailedEvent.cs
@@ -1,6 +1,6 @@
 using ECommerce.Shared.Infrastructure.EventBus;
 
-namespace Saga.Service.IntegrationEvents;
+namespace Saga.Service.Contracts.Integration.InboundEvents;
 
 public record ShipmentFailedEvent(
     Guid ShipmentId,
