+++ saga-microservice/Saga.Service/Program.cs
@@ -8,10 +8,10 @@ using ECommerce.Shared.OpenApi;
 using ECommerce.Shared.Qa;
 using OpenTelemetry.Metrics;
 using OpenTelemetry.Trace;
+using Saga.Service.Contracts.Integration.InboundEvents;
 using Saga.Service.Endpoints;
 using Saga.Service.Infrastructure.Data.EntityFramework;
 using Saga.Service.Infrastructure.Reaper;
-using Saga.Service.IntegrationEvents;
 using Saga.Service.IntegrationEvents.EventHandlers;
 using Saga.Service.Models;
 using Saga.Service.Observability;
