+++ saga-microservice/Saga.Tests/Domain/OrderCreatedEventHandlerTests.cs
@@ -5,9 +5,9 @@ using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging.Abstractions;
 using Microsoft.Extensions.Options;
+using Saga.Service.Contracts.Integration.InboundEvents;
 using Saga.Service.Infrastructure.Data.EntityFramework;
 using Saga.Service.Infrastructure.Reaper;
-using Saga.Service.IntegrationEvents;
 using Saga.Service.IntegrationEvents.EventHandlers;
 using Saga.Service.Models;
