+++ saga-microservice/Saga.Tests/EndToEnd/OrderSagaEndToEndTests.cs
@@ -4,8 +4,8 @@ using ECommerce.Shared.Infrastructure.Outbox;
 using ECommerce.Shared.IntegrationEvents.Commands;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
+using Saga.Service.Contracts.Integration.InboundEvents;
 using Saga.Service.Infrastructure.Data.EntityFramework;
-using Saga.Service.IntegrationEvents;
 using Saga.Service.IntegrationEvents.EventHandlers;
 using Saga.Service.Models;
